fix: resolve circular dep by colocating repo resolvers and repos
