+++ libs/backend/graphql/src/resolvers/app/app.resolver.ts
@@ -1,7 +1,1 @@
-import type { IResolvers } from '@graphql-tools/utils'
-
-export const appResolver: IResolvers = {
-  App: {
-    slug: ({ name }: { name: string }) => name.replace(/\s/g, '-'),
-  },
-}
+export const appSlug = ({ name }: { name: string }) => name.replace(/\s/g, '-')
