+++ libs/backend/graphql/src/resolvers/app/index.ts
@@ -1,1 +1,8 @@
-export * from './app.resolver'
+import type { IResolvers } from '@graphql-tools/utils'
+import { appSlug } from './app.resolver'
+
+export const appResolver: IResolvers = {
+  App: {
+    slug: appSlug,
+  },
+}
