+++ libs/backend/graphql/src/resolvers/domain/index.ts
@@ -1,13 +1,7 @@
 import type { IResolvers } from '@graphql-tools/utils'
-import { createDomains, deleteDomains, updateDomains } from './mutation'
 import { domainConfig, projectDomain } from './query'
 
 export const domainResolver: IResolvers = {
-  Mutation: {
-    createDomains,
-    deleteDomains,
-    updateDomains,
-  },
   Domain: {
     domainConfig,
     projectDomain,
