+++ libs/backend/graphql/src/resolvers/element/index.ts
@@ -1,13 +1,12 @@
-import { withReadTransactionResolver } from '@codelab/backend/infra/adapter/neo4j'
 import type { IResolvers } from '@graphql-tools/utils'
-import { elementDescendants } from './element.resolver'
+import { elementSlug } from './element.resolver'
+
+export * from './element.resolver'
 
 export const elementResolver: IResolvers = {
   Mutation: {},
   Query: {},
   Element: {
-    descendantElements: withReadTransactionResolver(elementDescendants),
+    slug: elementSlug,
   },
 }
-
-export * from './element.resolver'
