+++ libs/backend/graphql/src/resolvers/page/page.resolver.ts
@@ -1,9 +1,1 @@
-import type { IResolvers } from '@graphql-tools/utils'
-
-export const pageResolver: IResolvers = {
-  Mutation: {},
-  Query: {},
-  Page: {
-    slug: ({ name }: { name: string }) => name.replace(/\s/g, '-'),
-  },
-}
+export const pageSlug = ({ name }: { name: string }) => name.replace(/\s/g, '-')
