+++ libs/backend/graphql/tsconfig.lib.json
@@ -7,5 +7,9 @@
     "types": ["node"]
   },
   "exclude": ["jest.config.ts", "**/*.spec.ts", "**/*.test.ts"],
-  "include": ["**/*.ts"]
+  "include": [
+    "**/*.ts",
+    "../application/repo-resolvers/resolvers/domain/domain.error.ts",
+    "../application/repo-resolvers/resolvers/domain/domain.validation.ts"
+  ]
 }
