+++ libs/backend/infra/adapter/neo4j/src/infra/ogm.ts
@@ -1,4 +1,4 @@
-import { pureResolvers } from '@codelab/backend/pure-resolvers'
+import { resolvers } from '@codelab/backend/graphql'
 import type { OGM_TYPES } from '@codelab/shared/abstract/codegen'
 import { OGM } from '@neo4j/graphql-ogm'
 import { typeDefs } from '../schema'
@@ -11,7 +11,7 @@ export const getOgm = async () => {
   if (!ogm) {
     ogm = new OGM({
       typeDefs,
-      resolvers: pureResolvers,
+      resolvers,
       driver: getDriver(),
       config: {
         enableRegex: true,
