--- libs/backend/pure-resolvers/README.md
@@ -1,7 +1,0 @@
-# backend-pure-resolvers
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build backend-pure-resolvers` to build the library.
