--- libs/backend/pure-resolvers/src/index.ts
@@ -1,1 +1,0 @@
-export * from './pure-resolvers'
