--- libs/backend/pure-resolvers/src/pure-resolvers/element/element.resolver.ts
@@ -1,12 +1,0 @@
-import type { IResolvers } from '@graphql-tools/utils'
-
-export const elementSlug = ({ name }: { name: string }) =>
-  name.replace(/\s/g, '-')
-
-export const elementPureResolver: IResolvers = {
-  Mutation: {},
-  Query: {},
-  Element: {
-    slug: elementSlug,
-  },
-}
