--- libs/backend/pure-resolvers/src/pure-resolvers/element/index.ts
@@ -1,1 +1,0 @@
-export * from './element.resolver'
