--- libs/backend/pure-resolvers/src/pure-resolvers/index.ts
@@ -1,5 +1,0 @@
-import { mergeResolvers } from '@graphql-tools/merge'
-import type { IResolvers } from '@graphql-tools/utils'
-import { elementPureResolver } from './element'
-
-export const pureResolvers: IResolvers = mergeResolvers([elementPureResolver])
