+++ libs/backend/repo-resolvers/.eslintrc.json
@@ -5,7 +5,7 @@
     {
       "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
       "parserOptions": {
-        "project": ["libs/backend/pure-resolvers/tsconfig.*?.json"]
+        "project": ["libs/backend/repo-resolvers/tsconfig.*?.json"]
       },
       "rules": {}
     },
