+++ libs/backend/repo-resolvers/README.md
@@ -1,0 +1,7 @@
+# backend-repo-resolvers
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build backend-repo-resolvers` to build the library.
