+++ libs/backend/repo-resolvers/src/index.ts
@@ -1,0 +1,1 @@
+export * from './resolvers'
