+++ libs/backend/repo-resolvers/src/resolvers/atom/index.ts
@@ -2,11 +2,11 @@ import { withReadTransactionResolver } from '@codelab/backend/infra/adapter/neo4
 import type { IResolvers } from '@graphql-tools/utils'
 import { atoms } from './atom.resolver'
 
+export * from './atom.resolver'
+
 export const atomResolver: IResolvers = {
   Mutation: {},
   Query: {
     atoms: withReadTransactionResolver(atoms),
   },
 }
-
-export * from './atom.resolver'
