+++ libs/backend/repo-resolvers/src/resolvers/domain/index.ts
@@ -1,0 +1,10 @@
+import type { IResolvers } from '@graphql-tools/utils'
+import { createDomains, deleteDomains, updateDomains } from './mutation'
+
+export const domainResolver: IResolvers = {
+  Mutation: {
+    createDomains,
+    deleteDomains,
+    updateDomains,
+  },
+}
