+++ libs/backend/repo-resolvers/src/resolvers/domain/mutation/delete-domains.resolver.ts
