+++ libs/backend/repo-resolvers/src/resolvers/element/index.ts
@@ -1,0 +1,13 @@
+import { withReadTransactionResolver } from '@codelab/backend/infra/adapter/neo4j'
+import type { IResolvers } from '@graphql-tools/utils'
+import { elementDescendants } from './element.resolver'
+
+export const elementResolver: IResolvers = {
+  Mutation: {},
+  Query: {},
+  Element: {
+    descendantElements: withReadTransactionResolver(elementDescendants),
+  },
+}
+
+export * from './element.resolver'
