+++ libs/backend/repo-resolvers/src/resolvers/index.ts
@@ -1,0 +1,2 @@
+export * from './element'
+export * from './resolvers'
