+++ libs/backend/repo-resolvers/src/resolvers/tag/index.ts
@@ -1,6 +1,6 @@
 import { withReadTransactionResolver } from '@codelab/backend/infra/adapter/neo4j'
 import type { IResolvers } from '@graphql-tools/utils'
-import { tagDescendants } from './tag.resolvers'
+import { tagDescendants } from './tag.resolver'
 
 export const tagResolver: IResolvers = {
   Mutation: {},
